student_manager/requirements.txt
Package               Version
--------------------- -------
amqp                  5.0.6
asgiref               3.4.1
billiard              3.6.4.0
celery                5.1.2
click                 7.1.2
click-didyoumean      0.0.3
click-plugins         1.1.1
click-repl            0.2.0
Django                3.2.6
django-celery-results 2.2.0
django-rest-framework 0.1.0
djangorestframework   3.12.4
kombu                 5.1.0
pip                   21.1.3
prompt-toolkit        3.0.19
pytz                  2021.1
redis                 3.5.3
setuptools            56.0.0
six                   1.16.0
sqlparse              0.4.1
vine                  5.0.0
wcwidth               0.2.5
